fix/Corrindo interface dos componentes e trocando inicialização
