+++ qxcode-codigo/src/main/java/com/qxcode/Controller/ComponentController/IComponentController.java
@@ -1,10 +1,13 @@
-package com.qxcode.Controller;
+package com.qxcode.Controller.ComponentController;
 
+import com.qxcode.Model.Category;
+import com.qxcode.Model.IModel;
 import javafx.scene.input.MouseEvent;
 
 import java.io.IOException;
 
 public interface IComponentController {
+    public <T> void setModel(T model);
     void setInfo();
     void entrarDetalhes(MouseEvent mouseEvent) throws IOException;
 }
