+++ qxcode-codigo/src/main/resources/com/qxcode/View/components/categoryComponent.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane fx:id="cardCategory" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="240.0" prefWidth="350.0" xmlns="http://javafx.com/javafx/20.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.qxcode.Controller.CategoryComponent">
+<AnchorPane fx:id="cardCategory" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="240.0" prefWidth="350.0" xmlns="http://javafx.com/javafx/20.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.qxcode.Controller.ComponentController.CategoryComponent">
    <children>
       <Rectangle arcHeight="5.0" arcWidth="5.0" fill="WHITE" height="240.0" onMouseClicked="#entrarDetalhes" stroke="BLACK" strokeType="INSIDE" strokeWidth="0.0" style="-fx-background-color: #000000;" width="350.0" />
       <Label fx:id="categoryTitle" layoutX="14.0" layoutY="14.0" prefHeight="40.0" prefWidth="320.0" onMouseClicked="#entrarDetalhes" text="NOME DA CATEGORIA">
