+++ qxcode-codigo/src/main/resources/com/qxcode/View/telaCategory.fxml
@@ -10,7 +10,7 @@
 <?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane fx:id="telaCategory" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="832.0" prefWidth="1280.0" style="-fx-background-color: #0E1A25;" stylesheets="@../CSS/styleTelaCategory.css" xmlns="http://javafx.com/javafx/20.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.qxcode.Controller.TelaCategory">
+<AnchorPane fx:id="telaCategory" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="832.0" prefWidth="1280.0" style="-fx-background-color: #0E1A25;" stylesheets="@../CSS/styleTelaCategory.css" xmlns="http://javafx.com/javafx/20.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.qxcode.Controller.TelasController.TelaCategory">
 
     <children>
         <Pane fx:id="navBar" layoutX="80.0" layoutY="32.0" prefHeight="60.0" prefWidth="1120.0" />
