+++ qxcode-codigo/src/main/resources/com/qxcode/View/telaNewCategory.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.layout.Pane?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane fx:id="newCategory" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="832.0" prefWidth="1280.0" style="-fx-background-color: #0E1A25;" stylesheets="@../CSS/styleTelaQuestion.css" xmlns="http://javafx.com/javafx/20.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.qxcode.Controller.NewCategory">
+<AnchorPane fx:id="newCategory" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="832.0" prefWidth="1280.0" style="-fx-background-color: #0E1A25;" stylesheets="@../CSS/styleTelaQuestion.css" xmlns="http://javafx.com/javafx/20.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.qxcode.Controller.TelasController.NewCategory">
 
     <children>
         <Group layoutX="80.0" layoutY="124.0" AnchorPane.leftAnchor="80.0" AnchorPane.topAnchor="124.0">
