+++ qxcode-codigo/src/main/resources/com/qxcode/View/telaNewQuestion.fxml
@@ -13,7 +13,7 @@
 <?import javafx.scene.shape.Rectangle?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane fx:id="newQuestion" prefHeight="832.0" prefWidth="1280.0" style="-fx-background-color: #0E1A25;" stylesheets="@../CSS/styleTelaQuestion.css" xmlns="http://javafx.com/javafx/20.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.qxcode.Controller.NewQuestion">
+<AnchorPane fx:id="newQuestion" prefHeight="832.0" prefWidth="1280.0" style="-fx-background-color: #0E1A25;" stylesheets="@../CSS/styleTelaQuestion.css" xmlns="http://javafx.com/javafx/20.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.qxcode.Controller.TelasController.NewQuestion">
 
     <children>
         <Group layoutX="80.0" layoutY="124.0" AnchorPane.leftAnchor="80.0" AnchorPane.topAnchor="124.0">
